Task4 tested
Minor issues include duplicates in returned structure (i.e. stating the node C over and over again)
and not yet implementing the time frame threshold (currently using a counter of 10 for testing).

Next steps are to initiate with a random chain and then to change the constant in the MI weightings function
and analyse the results.